fix(lint): make the lint result platform-independent
CI runs golangci-lint on ubuntu; local runs on this machine are darwin. Build
tags mean the two see different code, so the gate reported 167 issues locally
and 171 on CI — with four findings invisible unless you ran GOOS=linux. A fix
verified only on darwin leaves CI red, and vice versa. That is a trap for
anyone working this backlog, so it is closed rather than worked around.

- trust/ports_common.go: pfTimeout moved into the darwin-only ports_macos.go.
  The untagged file's own comments say its constants live there 'so that tests
  on all platforms can reference the constant' — true for pfAnchorPath,
  launchDaemonPlistPath and launchDaemonPlistContent, which stay put, but
  nothing outside the darwin file ever referenced pfTimeout, so on linux it was
  simply dead.
- trust/coverage_g0t11_test.go: countOccurrences and indexOf moved into
  coverage_g0t11_darwin_test.go, the only file that uses them.
- trust/dns_linux.go: unchecked deferred f.Close, invisible from darwin.

Both platforms now report an identical 167. gofmt, build and vet clean under
GOOS=darwin and GOOS=linux; 4645 tests pass.
