+++ internal/trust/coverage_g0t11_test.go
@@ -200,34 +200,6 @@ func TestSetupLinux_NoSkipsViaSetupForceLinux(t *testing.T) {
 	})
 }
 
-// countOccurrences returns the number of times needle appears in haystack.
-func countOccurrences(haystack, needle string) int {
-	if needle == "" {
-		return 0
-	}
-	count := 0
-	pos := 0
-	for {
-		idx := indexOf(haystack[pos:], needle)
-		if idx < 0 {
-			break
-		}
-		count++
-		pos += idx + len(needle)
-	}
-	return count
-}
-
-// indexOf returns the first index of needle in haystack, or -1.
-func indexOf(haystack, needle string) int {
-	for i := 0; i+len(needle) <= len(haystack); i++ {
-		if haystack[i:i+len(needle)] == needle {
-			return i
-		}
-	}
-	return -1
-}
-
 // --- setupDarwin error-branch coverage via step seams ----------------------
 
 // stubError is a sentinel error type for tests.
